Testing of opening a ZIM file by file descriptor
The new unit-test `ZimFile.openByFD` was cloned from
`ZimFile.multipart`. It may not compile under windows because of the
direct usage of the open() system call (rather than `FS::openFile()`
which doesn't return a file descriptor).

The question is if the ZIM file should be opened by `FS::FD` instead
of `int fd`?